research/pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = "Research environment for bee monitoring system"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "dagshub>=0.6.9",
    "hydra-core>=1.3.2",
    "jupyter>=1.1.1",
    "kagglehub>=0.4.1",
    "mlflow>=3.11.1",
    "python-dotenv>=1.2.2",
    "ultralytics>=8.4.37",
    "optuna>=3.0.0",
    "seaborn>=0.13.2",
]
